game/widgets.py: Factor out list widgets from cyberdoc module.
